All the project has been restructured.
Now the project has a better structure:
- (~Controller) The content script interacts with the Gmail webpage reading and modifying the interface.
- (~View and model) The background has two files acting like event pages:
pgproject.js, that implements the logic of the application and interacts with openpgp.js (the library has been moved also to the background).
eventPage.js, that manages the messages between the different elements.

Functions has been adapted to work with this new structure, that makes more sense and fits better with the content scripts or event pages purposes.